Changed from bootstrap to css grid
Removed react-bootstrap 'grid' layout components in favor of grid
styling from css.
css grid makes a lot more sense for the type of display needed for the
set game and is much easier to manage - allowing for removal of a lot of
hacky code and general kruft.
It should also improve the consistency of the display across devices